+++ processor/src/test/resources/input/com/example/tracked/ObserveLowerPriorityTrackedModel.java
@@ -1,0 +1,19 @@
+package com.example.tracked;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.OnDepsChanged;
+import arez.annotations.Track;
+
+@ArezComponent
+public abstract class ObserveLowerPriorityTrackedModel
+{
+  @Track( observeLowerPriorityDependencies = true )
+  public void render( final long time, float someOtherParameter )
+  {
+  }
+
+  @OnDepsChanged
+  public void onRenderDepsChanged()
+  {
+  }
+}